add `StateProvider` for transactions
